separete mask creationt methods to ShadowDetecter and create new class, MaskCreater
